testing.yml: Fix commit messages
Note that you can access the commits and commit messages only when pushing or merging to the main branch, so for pull requests the github.event.commits is empty. It is anyway not needed, since we want to skip the step altogether for pull requests.

Test this anyway.

Exercise 11.16 - Skipping a commit for tagging and deployment